refactor: split abi tests
